Refactor tokenizing.
Decided against the Scanner and StringBuilder classes for the most part.
Now, for the most part, we just use the String methods replace, split,
and trim to tokenize the data.  Scanner is used for the more complicated
tokenization of the 4 digits and quote marks.

The pom file needed the maven compiler level in order to build the jar.