Model for Farm and FarmPoint.
Created the Farm object which is a list of list of FarmPoints.  Farm
also has the obvious method to calculate the fertile areas.  Needs to be
implemented of course.

FarmPoint is a simple point with values for which parcel it is in and
whether each point is fertile or not.