Added getInfertile area to rectangle.
In order to verify that all points have been set correctly in the walk,
we will need to know which points in each rectangle are infertile.  This
tells us that.  It is a simple test, but there is little reason to think
bigger rectangles would be wrong.

Hash code needed for point so you will find the point in the hashset.