Part of find intersections done.
The idea, was to see where an edge intersects with a rectangle.  This
doesn't handle the case where one point is in the rectangle and
intersects with the edge.


What I need next is to model the edges like sets of points.  Just loop
over the sets of points one at a time.

I want to check in an okish version first before I tear it apart.