Did a depth first search.
Instead of trying to play connect the dots, I did a simple queuing walk.
At anypoint, you look in all 8 directions. If it is fertile, unchecked
and in bounds, add it, and queue that point.  Repeat procedure with all
points.  When you visit a point, add it to the parcel total.

Now, I need to add a condition for WHICH parcel you are in.