Cleanup
Merge went badly.  Need to get  a merge conflict tool in SoureTree
and/or Eclipse.  Fixed it by hand.

I want to refactor this to have as a strategy pattern two different ways
to find the points.  First, I want to merge in what I have.  Then,  will
work on that.