Resuurected old code.
I want to do 2 algorithms here:  The O n^2 of walking all of the points
and finding islands.

The other one is finding where the rectangles intersect, determining
from that which areas are fertile and infertile, finding the areas, and
marking them appropriately.

The first step is to get back the rectangle intersection code.