First step in having a point matching algorithm.
First, we need to find the complete set of points.

IF the rectangles intersect with a rectangle OR the axis we need:  The
rectangle points, where they intersect with an axis, the intersection
points, AND a point on the axis just past the farthest left and right
points.

This is just the first trivial case of handling an empty list and a non
intersecting pair of rectangles.