Working on a skeleton for the find areas and use plots.
The first step is to find all the points were the rectangles intersect:

Axis.

Each other.

And, the point on the axis further right, left, above, and below the
furthest right, left, above, and below points.

The find intersection of rectangles code has to be very solid, so I
threw in a bunch of other test conditions.