Got the rectangle intersection finder working.
It is supposed to find where the rectangles intersect.

Find where the rectangles run the whole length or width of the plot,
there by creating a multiple fertile areas.

Include the origin if there is an intersection of any sort, including
axes.

Here is the tricky part:  If there is an intersection, we need to have
the following points on the axes:

	After the lower left point:  Point on axes left of and beneath.	
	After the upper left point:  Point on axes left of and above.
	After the lower right point:  Point on axes right of and below.
	After the upper right point:  Point on axes right of and above.
	
Added unit test to verfiy behavior.