Added sorted points to intersectoin finder.  Started Finding points to draw.
In order to find the points that we wish to draw to draw in the squares,
we need the list of points to be sorted, so the intersection finder
returns a TreeSet now.

We need a class to find where the next point should be, so that is
FindPointsToDraw.  Came up with test cases for the first method.  Will
implement after this commit.