First case of finding points to draw.
Implemeted the case of we are on an axis, find the next corner.  Go up
to that one.

Test cases also added for other cases, though the code is not written.

Added a y sorting comparator for points, as we will need to know the
next axis up in the y direction.

Also, allowed for corner points to be found and returned from the farm,
as we will need those as well as intersections.